test: move triggers + more tests
