+++ src/tests/triggers/mod.rs
@@ -1,0 +1,5 @@
+pub mod lidarr;
+pub mod notify;
+pub mod radarr;
+pub mod readarr;
+pub mod sonarr;
